requirements.txt
tqdm
h5py
timm
wandb
lpips
ninja
xatlas
pillow
einops
taichi
open3d
trimesh
lightning
matplotlib
hydra-core
blenderproc
tensorboard
scikit-image
torchmetrics
numpy<=1.27.0
point-cloud-utils
